refactor: Extract out `connect_relay.rs`
